Created the Player entity.
I changed the files for the Player entity - first of all, I the primary key is now an int, because it was hard and impractical to work with a String. I added the method which verifies if the username already exists in the database (see GetMapping). I also added the service layer.