Made the enter name screen functional.
This commit is pretty big - firstly, I changed the way screens are changed because it was very messy before and led to a lot of errors. Now it works without a problem. Secondly, I created the enter name screens for both the singleplayer and multiplayer mode, which I made functional. That is, you can enter a username (and if you don't provide one, you get a warning message). Then, when you provide a username and press play, the screen changes to the Singleplayer/Multiplayer. In the singleplayer mode, I also added a text on the top center with the username you provide.