Add CSS, load images and fix layout
Added and linked with CSS, load images for the back button and inside the tab pane. Each tab in the tab pane gets an image labelled 0-n loaded into it from the how2play folder. This way we can easily have as many how to play images as we want. I also fixed the layout of the how to play scene, as some stuff was out of bounds.