Add stylesheets to scenes, load images and fix remaining bugs
Everything should work now. Only the test fails because it uses outdated functions.

I added CSS loading to the initialize method of  MainCtrl, loaded in all the images  for the backButtons andfixed some remaining bugs.

Also, the contents of the initialize method of LeaderboardCtrl is commented out because it is broken. It was already broken, but just not used.

Also moved some stylesheets from resources/main to resources/stylesheets.