Changed methods in the file.
For now, new question appears once the player has chosen the answer for the current question. After he chose the answer, the app shows the correct answer and waits for 3 seconds and then changes the question. However, there are still problems with the correct answer, since the correct one is always the first activity. 
When the application finished running through all 20 questions, it exists.